Rename model package to data
