+++ src/seedu/addressbook/Main.java
@@ -1,15 +1,14 @@
 package seedu.addressbook;
 
-import seedu.addressbook.model.person.ReadOnlyPerson;
+import seedu.addressbook.data.person.ReadOnlyPerson;
 import seedu.addressbook.storage.StorageFile.*;
 
 import seedu.addressbook.commands.*;
-import seedu.addressbook.model.AddressBook;
+import seedu.addressbook.data.AddressBook;
 import seedu.addressbook.parser.Parser;
 import seedu.addressbook.storage.StorageFile;
 import seedu.addressbook.ui.TextUi;
 
-import java.io.*;
 import java.util.List;
 import java.util.Optional;
