+++ src/seedu/addressbook/commands/CommandResult.java
@@ -1,7 +1,6 @@
 package seedu.addressbook.commands;
 
-import seedu.addressbook.common.Utils;
-import seedu.addressbook.model.person.ReadOnlyPerson;
+import seedu.addressbook.data.person.ReadOnlyPerson;
 
 import java.util.List;
 import java.util.Optional;
