+++ src/seedu/addressbook/commands/FindPersonsByWordsInNameCommand.java
@@ -1,6 +1,6 @@
 package seedu.addressbook.commands;
 
-import seedu.addressbook.model.person.ReadOnlyPerson;
+import seedu.addressbook.data.person.ReadOnlyPerson;
 
 import java.util.*;
 import java.util.regex.Pattern;
