+++ src/seedu/addressbook/storage/jaxb/AdaptedTag.java
@@ -1,8 +1,8 @@
 package seedu.addressbook.storage.jaxb;
 
 import seedu.addressbook.common.Utils;
-import seedu.addressbook.model.exception.IllegalValueException;
-import seedu.addressbook.model.tag.Tag;
+import seedu.addressbook.data.exception.IllegalValueException;
+import seedu.addressbook.data.tag.Tag;
 
 import javax.xml.bind.annotation.XmlValue;
 
@@ -32,7 +32,7 @@ public class AdaptedTag {
      * Checks whether any required element is missing.
      *
      * JAXB does not enforce (required = true) without a given XML schema.
-     * Since we do most of our validation using the model class constructors, the only extra logic we need
+     * Since we do most of our validation using the data class constructors, the only extra logic we need
      * is to ensure that every xml element in the document is present. JAXB sets missing elements as null,
      * so we check for that.
      */
@@ -41,7 +41,7 @@ public class AdaptedTag {
     }
 
     /**
-     * Converts this jaxb-friendly adapted tag object into the model's Tag object.
+     * Converts this jaxb-friendly adapted tag object into the Tag object.
      *
      * @throws IllegalValueException if there were any data constraints violated in the adapted person
      */
